COC-206-refactor(#149): SpaceCard 리팩토링
